\name{FAVErtools-package}
\alias{FAVErtools-package}
\alias{FAVErtools}
\docType{package}
\title{
This is a set of tools for working with the output of FAVE.
}
\description{
Right now, FAVErtools only contains scripts for reading in plotnik files.
}
\details{
\tabular{ll}{
Package: \tab FAVE_rtools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-27\cr
License: \tab MIT\cr
}
}
\author{
Josef Fruehwald <jofrhwld@gmail.com>
}
\keyword{ package }
