\name{plt_vowels}
\alias{plt_vowels}
\title{Translate numerical plotnik codes to readable codes}
\usage{
  plt_vowels(x)
}
\arguments{
  \item{x}{a vector of numeric vowel codes}
}
\description{
  This will translate numerical plotnik codes to a readable
  code. Note, this only operates over the part of the
  numeric code preceding the decimal.
}

